requirements.txt
opencv-python
numpy
ultralytics
pyserial
python-dotenv
torch
torchvision
torchaudio
